requirements.txt
requests==2.32.3
pandas==2.2.3
google-cloud-bigquery==3.27.0
dbt-core==1.9.0
dbt-bigquery==1.9.0
prefect>=3.0.0
python-dotenv==1.0.1
sqlfluff==3.2.5
sqlfluff-templater-dbt==3.2.5